CIV-5984 fix bug for common hearing requirements
